Moved config infrastructure to parent folder
