+++ test/express/integration/tables/data.tests.js
@@ -6,9 +6,9 @@ var expect = require('chai')
         .expect,
     supertest = require('supertest-as-promised'),
     express = require('express'),
-    mobileApps = require('../../infrastructure/mobileApps'),
+    mobileApps = require('../../../infrastructure/mobileApps'),
     data = require('../../../../src/data/mssql'),
-    config = require('../../infrastructure/config'),
+    config = require('../../../infrastructure/config'),
 
     app, mobileApp;
