+++ test/express/integration/tables/dynamicSchema.tests.js
@@ -4,9 +4,9 @@
 var expect = require('chai').use(require('chai-subset')).expect,
     supertest = require('supertest-as-promised'),
     express = require('express'),
-    mobileApps = require('../../infrastructure/mobileApps'),
+    mobileApps = require('../../../infrastructure/mobileApps'),
     data = require('../../../../src/data/mssql'),
-    config = require('../../infrastructure/config'),
+    config = require('../../../infrastructure/config'),
     promises = require('../../../../src/utilities/promises'),
 
     app, mobileApp;
