+++ test/express/integration/tables/initialize.tests.js
@@ -4,8 +4,8 @@
 ﻿var expect = require('chai').expect,
     supertest = require('supertest-as-promised'),
     express = require('express'),
-    mobileApps = require('../../infrastructure/mobileApps'),
-    config = require('../../infrastructure/config'),
+    mobileApps = require('../../../infrastructure/mobileApps'),
+    config = require('../../../infrastructure/config'),
     data = require('../../../../src/data/mssql'),
     promises = require('../../../../src/utilities/promises'),
